Make image storage cleanup recoverable
